lnd: introduce the ChannelNotifier.
This commit introduces the channel notifier which is a central source of active, inactive, and closed channel events.

This notifier was originally intended to be used by the `SubscribeChannels` streaming RPC call, but can be used by any subsystem that needs to be notified on a channel becoming active, inactive or closed.

It may also be extended in the future to support other types of notifications.