rpcserver: UpdateChanPolicy can specify max htlc
In this commit, we enable callers of UpdateChannelPolicy to
specify their desired max HTLC forwarding policy for one or
multiple channels.

We also add checks for the specified value: the value can't
be above the max pending msats for the channel, nor less
than the minimum supported HTLC for the channel or an
error is returned.