discovery/gossiper+rpcserver: handle new max htlc
In this commit, we ensure that when a caller of
UpdateChannelPolicy specifies a new max htlc, the value is
persisted to disk and broadcast to the network.

The gossiper's`updateChannel` handles writing new policy
updates to disk as well as retrieving stale edge policies for
rebroadcast. Stale policies may not have a max htlc set,
so this commit includes a check to ensure the message
flag is set appropriately.